--- src/main/java/com/tolive/ddubeok2/ConfigController.java
@@ -1,19 +1,0 @@
-package com.tolive.ddubeok2;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.ModelAndView;
-
-@Controller
-public class ConfigController {
-	
-	@RequestMapping(value="/home.do")
-	public ModelAndView indexRequest() {
-		
-		ModelAndView modelAndView = new ModelAndView();
-		modelAndView.setViewName("home");
-		
-		return modelAndView;
-	}
-}
-
